requirements.txt
wxauto==3.9.11.17.5
schedule==1.2.2
itchat-uos==1.5.0.dev0
requests==2.31.0 